Fix Windows build and runtime errors.
Remove usage of C++11 unions from MatrixProcessor. The functionality
is the same. Visual Studio does not yet support C++11 unions.

Properly initialize YRange in BulkSplineEvaluator.

Add Windows build artifacts to gitignore.

Add proper casts between MotiveTime and float for the spline
evaluator.

Remove extraneous 'typename' keywords that Visual Studio didn't like.

Remove reference to log2 from Visual Studio build. It's not
supported.

Tested: Motive and PieNoon build on Mac, Linux, Android, and Windows.

+++ docs/samples/smooth1f.cpp
@@ -79,7 +79,8 @@ int main() {
     // Additionally, we can also access facing_angle.Velocity() for the angular
     // velocity.
     const Angle facing_angle_at_time_t(facing_angle.Value());
-    points.push_back(vec2(t, facing_angle_at_time_t.ToDegrees()));
+    points.push_back(
+        vec2(static_cast<float>(t), facing_angle_at_time_t.ToDegrees()));
   }
 
   printf("\n%s", Graph2DPoints(&points[0], points.size()).c_str());
